+++ bin/config/config.ts
@@ -2,6 +2,8 @@ import path from 'node:path';
 import { isExists } from '@stlib/utils';
 import fs from 'fs';
 import YAML from 'yaml';
+import { ConfigException } from '../../lib/exceptions';
+import { Cli } from '../cli';
 
 export type ConfigType = {
   pattern?: string;
@@ -31,6 +33,16 @@ export class Config {
       'stest.config.ts',
     ];
 
+    const configOption = Cli.getOptions('config');
+
+    if (configOption && typeof configOption === 'string') {
+      await this.setConfigPathIterator([configOption]);
+    } else {
+      await this.setConfigPathIterator(configFileNames);
+    }
+  }
+
+  private static async setConfigPathIterator(configFileNames: string[]) {
     for (const fileName of configFileNames) {
       const configPath = path.join(this.projectPath, fileName);
       if (await isExists(configPath)) {
@@ -68,7 +80,9 @@ export class Config {
       const importedConfig = await import(filePath);
       return importedConfig.default || importedConfig;
     } catch (error) {
-      console.error(`Failed to load configuration from ${filePath}:`, error);
+      console.error(
+        new ConfigException(`Failed to load configuration from ${filePath}`),
+      );
       return undefined;
     }
   }