+++ bin/config/index.ts
@@ -1,1 +1,3 @@
-export { Config, ConfigType as StestConfig } from './config';
+export { Config, ConfigType } from './config';
+export { ConfigType as StestConfig } from './config'; // for package types
+export * from './config.init';