+++ bin/spinner/spinner.ts
@@ -1,9 +1,13 @@
 import { createSpinner } from 'nanospinner';
 
-export const spinner = createSpinner('Loading test files');
+export const spinner = createSpinner();
 
-export const spinnerWrapper = async (callback: Promise<any> | any) => {
-  spinner.start();
-  await callback();
-  spinner.success();
+export const spinnerWrapper = async (
+  callback: Promise<any> | any,
+  args: any[],
+  message?: string,
+) => {
+  spinner.start({ text: message });
+  await callback(...args);
+  spinner.success({ text: message });
 };