+++ lib/types/index.ts
@@ -1,2 +1,3 @@
 export * from './anyFunction.type';
 export * from './methodName.type';
+export * from './lifecycle.type';